requirements.txt
colorama
audioop-lts
python-dotenv
discord.py-self